requirements.txt
coverage==4.3.4
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gevent==1.1.2
Flask-Cors==3.0.0
Flask-SQLAlchemy==2.1
pymysql==0.7.10
PyJWT==1.4.2
Flask-Cache
